requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.97.0
djangorestframework-simplejwt==5.5.1
Pillow==12.2.0
google-genai==1.73.1
anyio==4.13.0
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.4.22
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==6.0.4
django-cors-headers==4.9.0
django_celery_results==2.6.0
djangorestframework==3.17.1
docstring_parser==0.18.0
flower==2.0.1
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
idna==3.13
jiter==0.14.0
kombu==5.6.2
lxml==6.1.0
packaging==26.2
prometheus_client==0.25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
PyMuPDF==1.27.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
pytz==2026.1.post1
redis==7.4.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
